requirements.txt
annotated-types==0.6.0
anyio==4.1.0
argcomplete==2.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
autocommand==2.2.2
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
btrfsutil==6.6.2
build==1.0.3
cairocffi==1.6.1
catboost==1.2.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.0
contourpy==1.2.0
cryptography==41.0.6
cycler==0.12.1
dbus-next==0.2.3
dbus-python==1.3.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
dnspython==2.4.2
evdev==1.6.1
executing==2.0.0
fastjsonschema==2.19.0
fonttools==4.44.3
fqdn==1.5.1
googlesearch-python==1.2.3
graphviz==0.20.1
idna==3.4
importlib-metadata==5.0.0
inflect==7.0.0
installer==0.7.0
ipykernel==6.27.1
ipython==8.17.2
isoduration==20.11.0
jaraco.context==4.3.0
jaraco.functools==3.9.0
jaraco.text==3.11.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mistune==3.0.2
moddb==0.10.0
more-itertools==10.1.0
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.7
netsnmp-python==1.0a1
notebook==7.0.6
notebook_shim==0.2.3
numpy==1.26.2
ordered-set==4.1.0
overrides==7.4.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
Pillow==10.1.0
pipx==1.2.1
platformdirs==4.0.0
plotly==5.18.0
ply==3.11
pooch==1.8.0
prometheus-client==0.19.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.25.1
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.16.1
PyGObject==3.46.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyrate-limiter==3.1.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3.post1
pywayland==0.4.17
pywlroots==0.16.6
PyYAML==6.0.1
pyzmq==25.1.1
qtile==0.23.0
qtile-extras==0.23.0
referencing==0.31.1
Reflector==2023.6.28.0.36.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.13.2
scikit-learn==1.3.2
SciPy==1.11.4
seaborn==0.12.2
Send2Trash==1.8.2
six==1.16.0
smbus==1.1
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.3
stack-data==0.6.3
tenacity==8.2.3
terminado==0.18.0
threadpoolctl==3.2.0
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
traitlets==5.14.0
trove-classifiers==2023.11.23
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
uri-template==1.3.0
urllib3==1.26.18
userpath==1.9.1
validate-pyproject==0.13.post1.dev0+gb752273.d20230520
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
whois==0.9.27
xcffib==1.4.0
xgboost==2.0.2
xkbcommon==0.8
zipp==3.17.0
